ironic: Move factory code to separate package
